Refactor more ElasticSearch jobs/tests into the ElasticSearch namespace
Lets move ElasticSearchIndexDelete.php when #300 is merged
